feat: improve auth service, add JWT support

+++ auth/model/refresh_token.go
@@ -3,6 +3,6 @@ package model
 type RefreshToken struct {
 	Token   string `json:"token"`
 	UserId  int    `json:"user_id"`
-	Expires uint   `json:"expires"`
+	Expires int64   `json:"expires"`
 	ID 		int    `json:"id"`	
 }
